% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_sse.R
\name{gg_sse}
\alias{gg_sse}
\title{Add Secondary Structure Annotation to a ggplot Matrix}
\usage{
gg_sse(x, min = -5, max = 0, helix.col = "gray20", sheet.col = "gray80",
  sse.border = "black", side = c(1, 2))
}
\arguments{
\item{x}{A bio3d object with secondary structure information. Most commonly
this is a pdb or pdbs class object as obtained from the bio3d functions
read.pdb and read.fasta.pdbs. However, a vector of sse elements as obtained
from the function pdb2sse is also valid input. The later is useful for more
complicated multi-chain input structures. See examples.}

\item{min}{Currently the bottom most coordinate  for annotation rectangles. 
More negative values will result in larger width rectangles. Eventually
we will change this to introduce a single input for specifying the relative
width of sse annotation rectangles.}

\item{max}{the top most plot coordinate for annotation rectangles.}

\item{helix.col}{The fill colors for rectangles representing alpha helices.}

\item{sheet.col}{The fill colors for rectangles representing beta strands.}

\item{sse.border}{The stroke color for rectangle borders.}

\item{side}{A numeric vector with values 1 to 4 specifying he side of the 
plot where annotations will be drawn. Note sides 3 and 4 are not set}
}
\value{
A list of annotation layers for adding to a ggplot object.
}
\description{
Add marginal secondary structure elements (SSEs) annotation layer to the
  current ggplot object.
}
\note{
There is lots to clean up and improve in this function. This includes 
  making on one call to annotate() for all rectangles rather than the four  
  calls per input type we do currently. Also we need a side=c(1:4) option  
  so we can add this for 1d plots and 2d plots (with the later providing any
   combination of annotated axis). 

  Two forms of secondary structure annotation are available: so
  called ‘classic’ and ‘fancy’. The former draws marginal rectangles
  and has been available within Bio3D from version 0.1. The later
  draws more ‘fancy’ (and distracting) 3D like helices and arrowed strands.
  Also ToDo: add side options to allow drawing on sides 1 to 4
}
\examples{
 pdb <- read.pdb( "5p21" )
 k <- dm(pdb, inds="calpha", mask.lower=FALSE)
 gg.mat(k)

 gg.mat(k) + gg_sse(pdb) ## Add secondary structure from pdb

}

