% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_sse.R
\name{gg_sse}
\alias{gg_sse}
\title{Add Secondary Structure Annotation to a ggplot Matrix}
\usage{
gg_sse(x, min = -5, max = 0, helix.col = "gray20", sheet.col = "gray80",
  sse.border = "black")
}
\arguments{
\item{x}{A bio3d object with secondary structure information. Most commonely
this is a pdb or pdbs class object as obtained from the bio3d functions
read.pdb and read.fasta.pdbs. However, a vector of sse elements as obtained
from the function pdb2sse is also valid input. The later is usefull for more
compliacted multi-chain input structures. See examples.}

\item{min}{Currently the bottom coord for annotation rectangles. More negative
values will result in larger width rectangles. Eventually this will be the
relative width of sse annotation rectangles.}

\item{max}{the top coord of annotation rectangles.}

\item{helix.col}{The fill colors for rectangles representing alpha helices.}

\item{sheet.col}{The fill colors for rectangles representing beta strands.}

\item{sse.border}{The stroke color for rectangle borders.}
}
\value{
A list of annotation layers for adding to a ggplot object.
}
\description{
Add marginal secondary structure elements (SSEs) annotation layer to the
  current ggplot object.
}
\note{
Two forms of secondary structure annotation are available: so
  called ‘classic’ and ‘fancy’. The former draws marginal rectangles
  and has been available within Bio3D from version 0.1. The later
  draws more ‘fancy’ (and distracting) 3D like helices and arrowed strands.
  Also ToDo: add side options to allow drawing on sides 1 to 4
}
\examples{
 pdb <- bio3d::read.pdb( "5p21" )
 k <- bio3d::dm(pdb, inds="calpha", mask.lower=FALSE)
 ggplot.dmat(k)

 ggplot.dmat(k) + gg_sse(pdb) ## Add secondary structure from pdb
}

