% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.dmat.R
\name{ggplot.dmat}
\alias{ggplot.dmat}
\title{ggplot bio3d matrix data}
\usage{
ggplot.dmat(dm)
}
\arguments{
\item{dm}{A numeric matrix such as a distance matrix from the bio3d functions
\code{dm}, \code{cmap}, \code{dccm} etc. Could write these as \code{\link[bio3d]{dm}},
\code{\link[bio3d]{cmap}}, \code{\link[bio3d]{dccm}}, \code{pca.array.loadings},
etc.}
}
\value{
A ggplot object for plotting with \code{print.ggplot}.
}
\description{
\code{ggplot.dmat} returns a ggplot version of plot.dmat() NxN matrix plot.
}
\details{
This is a simple first go at getting our usual bio3d matrices into a ggplot
  friendly format. This works with utility functions for adding annotations
  such as secondary structure, domain boundarys highlight regions etc.
}
\note{
Still much to be improved here, including adding extra args via dots.
}
\examples{
 pdb <- bio3d::read.pdb( "5p21" )
 k <- bio3d::dm(pdb, inds="calpha", mask.lower=FALSE)
 ggplot.dmat(k)

 ggplot.dmat(k) + gg_sse(pdb) ## Add secondary structure from pdb

 ggplot(k) + theme_grey() +   ## The ggplot2 gray background theme
   scale_x_continuous(expand = c(0, 0)) +
   scale_y_continuous(expand = c(0, 0))  ## start axis at zero
}
\seealso{
The bio3d base, grid and lattice ploting functions for various matrix class objects.
  This includes \code{\link[bio3d]{plot.dmat}}, \code{\link[bio3d]{plot.dccm}},
  \code{\link[bio3d]{plot.cmap}}, etc.
}

